# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WWHS Digital Campus API",
    "version": "0.1.0",
    "description": "Comprehensive REST API for WWHS Digital Campus covering authentication, academics, attendance, LMS, exams, timetable, notifications, library, finance, CMS, and AI assistant."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/documents": {
      "get": {
        "summary": "List document records with optional filters",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "docType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentProfileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "staffProfileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "academicYearId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document records retrieved successfully"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      },
      "post": {
        "summary": "Create a new document record",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "docType": {
                    "type": "string",
                    "minLength": 1
                  },
                  "studentProfileId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "staffProfileId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "academicYearId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "metadataJson": {
                    "type": "object"
                  }
                },
                "required": [
                  "docType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document record created successfully"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          }
        }
      }
    },
    "/api/documents/payload/{docType}/{referenceId}": {
      "get": {
        "summary": "Generate printable document payload for result cards, receipts, certificates, etc.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "docType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Printable document payload generated successfully"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error400": {
        "description": "Validation error or bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error401": {
        "description": "Unauthenticated or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Error404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}